% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histoslider.R
\name{update_histoslider}
\alias{update_histoslider}
\title{Update a histoslider}
\usage{
update_histoslider(
  id,
  label = NULL,
  values = NULL,
  start = NULL,
  end = NULL,
  width = NULL,
  height = NULL,
  breaks = "Sturges",
  options = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\description{
Update a histoslider
}
